requirements.txt
configparser==5.2.0
fastapi==0.75.1
mysql-connector-python==8.0.27
pathlib==1.0.1
pydantic==1.9.0
starlette==0.17.1
typing-extensions==4.0.1
uvicorn==0.17.6